+++ ProjectPlans/ProjectPlan_u7582024/Project_Putri/streamlit_app.py
@@ -18,7 +18,6 @@ import geopandas as gpd
 import matplotlib.pyplot as plt
 import folium
 from streamlit_folium import st_folium
-import branca
 import branca.colormap as cm
 import os
 import sys
